scrollview and refactor components structure
