--- src/components/form/index.ts
@@ -1,2 +1,0 @@
-export { default as Input } from './Input'
-export { default as DateInput } from './DateInput'
