+++ src/components/index.ts
@@ -1,0 +1,27 @@
+//ui
+export { default as Accordion } from './ui/Accordion'
+export { default as Body } from './ui/Body'
+export { default as Box } from './ui/Box'
+export { default as Center } from './ui/Center'
+export { default as Divider } from './ui/Divider'
+export { default as ErrorText } from './ui/ErrorText'
+export { default as H1 } from './ui/H1'
+export { default as H2 } from './ui/H2'
+export { default as H3 } from './ui/H3'
+export { default as H4 } from './ui/H4'
+export { default as H5 } from './ui/H5'
+export { default as H6 } from './ui/H6'
+export { default as HStack } from './ui/HStack'
+export { default as Icon } from './ui/Icon'
+export { default as Label } from './ui/Label'
+export { default as Link } from './ui/Link'
+export { default as PrimaryButton } from './ui/PrimaryButton'
+export { default as ScrollView } from './ui/ScrollView'
+export { default as SecondaryButton } from './ui/SecondaryButton'
+export { default as VStack } from './ui/VStack'
+
+//form
+export { default as DateInput } from './form/DateInput'
+export { default as Input } from './form/Input'
+
+//dialogs
