+++ src/components/ui/ScrollView/index.tsx
@@ -1,0 +1,17 @@
+import React, { ComponentProps, FC } from 'react'
+import {
+  Keyboard,
+  ScrollView as RNScrollView,
+  TouchableWithoutFeedback,
+} from 'react-native'
+
+const ScrollView: FC<ComponentProps<typeof RNScrollView>> = ({
+  children,
+  ...rest
+}) => (
+  <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
+    <RNScrollView {...rest}>{children}</RNScrollView>
+  </TouchableWithoutFeedback>
+)
+
+export default ScrollView
