--- src/components/ui/index.ts
@@ -1,19 +1,0 @@
-export { default as Accordion } from './Accordion'
-export { default as Body } from './Body'
-export { default as Box } from './Box'
-export { default as Center } from './Center'
-export { default as Divider } from './Divider'
-export { default as ErrorText } from './ErrorText'
-export { default as H1 } from './H1'
-export { default as H2 } from './H2'
-export { default as H3 } from './H3'
-export { default as H4 } from './H4'
-export { default as H5 } from './H5'
-export { default as H6 } from './H6'
-export { default as HStack } from './HStack'
-export { default as Icon } from './Icon'
-export { default as Label } from './Label'
-export { default as Link } from './Link'
-export { default as PrimaryButton } from './PrimaryButton'
-export { default as SecondaryButton } from './SecondaryButton'
-export { default as VStack } from './VStack'
