+++ src/screens/FormElementScreen/FormElementScreen.style.ts
@@ -4,7 +4,10 @@ import { styleManager } from '@/libs'
 
 export const formElementScreenStyles = styleManager.createStyleSheet(theme => ({
   container: {
+    flex: 1,
+    padding: theme.margins.xl,
+  },
+  wrapper: {
     gap: scale(10),
-    margin: theme.margins.xl,
   },
 }))
