+++ src/validations/formElementSchema.ts
@@ -2,8 +2,8 @@ import { z } from 'zod'
 
 const formElementSchema = () => {
   return z.object({
-    input: z.string(),
-    dateInput: z.string(),
+    input: z.string().min(5).max(10),
+    dateInput: z.string().min(5).max(10),
   })
 }
